% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashrafi2_2018.R
\name{ashrafi2_2018}
\alias{ashrafi2_2018}
\title{Ashrafi II model for fitting thermal performance curves}
\usage{
ashrafi2_2018(temp, a, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{dimensionless parameter}

\item{b}{dimensionless parameter}

\item{c}{dimensionless parameter}
}
\value{
a numeric vector of rate values based on the temperatures and parameter values provided to the function
}
\description{
Ashrafi II model for fitting thermal performance curves
}
\details{
Equation:
\deqn{rate=a + b \cdot temp^{\frac{3}{2}} + c \cdot temp^{2}}{%
rate = a + b.temp^(3/2) + c.temp^2}

Start values in \code{get_start_vals} are derived from the data or sensible values from the literature.

Limits in \code{get_lower_lims} and \code{get_upper_lims} are derived from the data or based extreme values that are unlikely to occur in ecological settings.
}
\note{
Generally we found this model easy to fit.
}
\examples{
# load in ggplot
library(ggplot2)

# subset for the first TPC curve
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'ashrafi2_2018')
# fit model
mod <- nls.multstart::nls_multstart(rate~ashrafi2_2018(temp = temp, a, b, c),
data = d,
iter = c(4,4,4),
start_lower = start_vals - 10,
start_upper = start_vals + 10,
lower = get_lower_lims(d$temp, d$rate, model_name = 'ashrafi2_2018'),
upper = get_upper_lims(d$temp, d$rate, model_name = 'ashrafi2_2018'),
supp_errors = 'Y',
convergence_count = FALSE)

# look at model fit
summary(mod)

# get predictions
preds <- data.frame(temp = seq(min(d$temp), max(d$temp), length.out = 100))
preds <- broom::augment(mod, newdata = preds)

# plot
ggplot(preds) +
geom_point(aes(temp, rate), d) +
geom_line(aes(temp, .fitted), col = 'blue') +
theme_bw()
}
\references{
Ashrafi, R. et al. Broad thermal tolerance is negatively correlated with virulence in an opportunistic bacterial pathogen. Evolutionary Applications 11, 1700–1714 (2018).
}
\author{
Francis Windram
}
\concept{model}
