% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atkin_2005.R
\name{atkin_2005}
\alias{atkin_2005}
\title{Atkin model for fitting thermal performance curves}
\usage{
atkin_2005(temp, r0, a, b)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{r0}{scaling parameter, the minimum trait value}

\item{a}{arbitrary scaling parameter}

\item{b}{arbitrary scaling parameter}
}
\value{
a numeric vector of rate values based on the temperatures and parameter values provided to the function
}
\description{
Atkin model for fitting thermal performance curves
}
\details{
Equation:
\deqn{rate = B_0 \cdot (a - b \cdot T)^{\frac{T}{10}}}{%
rate = B0.(a-b.T)^(T/10)}

Start values in \code{get_start_vals} are derived from the data or sensible values from the literature.

Limits in \code{get_lower_lims} and \code{get_upper_lims} are derived from the data or based extreme values that are unlikely to occur in ecological settings.
}
\note{
Generally we found this model easy to fit.
}
\examples{
# load in ggplot
library(ggplot2)

# subset for the first TPC curve
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'atkin_2005')
# fit model
mod <- nls.multstart::nls_multstart(rate~atkin_2005(temp = temp, r0, a, b),
data = d,
iter = 200,
start_lower = start_vals - 10,
start_upper = start_vals + 10,
lower = get_lower_lims(d$temp, d$rate, model_name = 'atkin_2005'),
upper = get_upper_lims(d$temp, d$rate, model_name = 'atkin_2005'),
supp_errors = 'Y',
convergence_count = FALSE)

# look at model fit
summary(mod)

# get predictions
preds <- data.frame(temp = seq(min(d$temp), max(d$temp), length.out = 100))
preds <- broom::augment(mod, newdata = preds)

# plot
ggplot(preds) +
geom_point(aes(temp, rate), d) +
geom_line(aes(temp, .fitted), col = 'blue') +
theme_bw()
}
\references{
Atkin, OK, Bruhn D, Tjoelker MG. Response of Plant Respiration to Changes in Temperature: Mechanisms and Consequences of Variations in Q10 Values and Acclimation. In Plant Respiration. 2005.
}
\author{
Francis Windram
}
