% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_2012.R
\name{beta_2012}
\alias{beta_2012}
\title{Beta model for TPC}
\usage{
beta_2012(temp, a, b, c, d, e)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{dimensionless parameter}

\item{b}{dimensionless parameter}

\item{c}{dimensionless parameter}

\item{d}{dimensionless parameter}

\item{e}{dimensionless parameter}
}
\description{
Beta model for TPC
}
\references{
Niehaus, Amanda C., et al. "Predicting the physiological performance of ectotherms in fluctuating thermal environments." Journal of Experimental Biology 215.4 (2012): 694-701.
}
\author{
Daniel Padfield
}
