% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betatypesimplified_2008.R
\name{betatypesimplified_2008}
\alias{betatypesimplified_2008}
\title{Simplified Beta-type model for fitting thermal performance curves}
\usage{
betatypesimplified_2008(temp, rho, alpha, beta)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{rho}{dimensionless parameter}

\item{alpha}{dimensionless parameter}

\item{beta}{dimensionless parameter}
}
\value{
a numeric vector of rate values based on the temperatures and parameter values provided to the function
}
\description{
Simplified Beta-type model for fitting thermal performance curves
}
\details{
Equation:
\deqn{rate = \rho \cdot \left(a - \frac{T}{10}\right) \cdot \left(\frac{T}{10}\right)^b}{%
rate = rho . (a - (T/10) . (T/10)^b)}

Start values in \code{get_start_vals} are derived from the data or sensible values from the literature.

Limits in \code{get_lower_lims} and \code{get_upper_lims} are derived from the data or based extreme values that are unlikely to occur in ecological settings.
}
\note{
Generally we found this model easy to fit.
}
\examples{
\donttest{
# load in ggplot
library(ggplot2)

# subset for the first TPC curve
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'betatypesimplified_2008')
# fit model
mod <- nls.multstart::nls_multstart(rate~betatypesimplified_2008(temp = temp, rho, alpha, beta),
data = d,
iter = c(7,7,7),
start_lower = start_vals - 10,
start_upper = start_vals + 10,
lower = get_lower_lims(d$temp, d$rate, model_name = 'betatypesimplified_2008'),
upper = get_upper_lims(d$temp, d$rate, model_name = 'betatypesimplified_2008'),
supp_errors = 'Y',
convergence_count = FALSE)

# look at model fit
summary(mod)

# get predictions
preds <- data.frame(temp = seq(min(d$temp), max(d$temp), length.out = 100))
preds <- broom::augment(mod, newdata = preds)

# plot
ggplot(preds) +
geom_point(aes(temp, rate), d) +
geom_line(aes(temp, .fitted), col = 'blue') +
theme_bw()
}
}
\references{
Damos, P. & Savopoulou-Soultani, M. Temperature-dependent bionomics and modeling of Anarsia lineatella (Lepidoptera: Gelechiidae) in the laboratory. J. Econ. Entomol. 101, 1557–1567 (2008).
}
\author{
Francis Windram
}
