% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briereextended_2021.R
\name{briereextended_2021}
\alias{briereextended_2021}
\title{Extended Brière model for fitting thermal performance curves}
\usage{
briereextended_2021(temp, tmin, tmax, a, b, d)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{tmin}{low temperature (ºC) at which rates become negative}

\item{tmax}{high temperature (ºC) at which rates become negative}

\item{a}{scale parameter to adjust maximum rate of the curve}

\item{b}{shape parameter to adjust the asymmetry of the curve}

\item{d}{shape parameter to adjust the asymmetry of the curve}
}
\value{
a numeric vector of rate values based on the temperatures and parameter values provided to the function
}
\description{
Extended Brière model for fitting thermal performance curves
}
\details{
Equation:
\deqn{rate=a \cdot temp \cdot (temp - t_{min})^b \cdot (t_{max} - temp)^d}{%
rate = a.temp.(temp - tmin)^b.(tmax - temp)^d}

Start values in \code{get_start_vals} are derived from the data or sensible values from the literature.

Limits in \code{get_lower_lims} and \code{get_upper_lims} are derived from the data or based extreme values that are unlikely to occur in ecological settings.
}
\note{
Generally we found this model easy to fit.
}
\examples{
# load in ggplot
library(ggplot2)

# subset for the first TPC curve
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'briereextended_2021')
# fit model
mod <- nls.multstart::nls_multstart(rate~briereextended_2021(temp = temp, tmin, tmax, a, b, d),
data = d,
iter = c(4,4,4,4,4),
start_lower = start_vals - 10,
start_upper = start_vals + 10,
lower = get_lower_lims(d$temp, d$rate, model_name = 'briereextended_2021'),
upper = get_upper_lims(d$temp, d$rate, model_name = 'briereextended_2021'),
supp_errors = 'Y',
convergence_count = FALSE)

# look at model fit
summary(mod)

# get predictions
preds <- data.frame(temp = seq(min(d$temp), max(d$temp), length.out = 100))
preds <- broom::augment(mod, newdata = preds)

# plot
ggplot(preds) +
geom_point(aes(temp, rate), d) +
geom_line(aes(temp, .fitted), col = 'blue') +
theme_bw()
}
\references{
Cruz-Loya, M. et al. Antibiotics shift the temperature response curve of Escherichia coli growth. mSystems 6, e00228–21 (2021).
}
\concept{model}
