% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delong_2017.R
\name{delong_2017}
\alias{delong_2017}
\title{deLong enzyme-assisted Arrhenius model}
\usage{
delong_2017(temp_k, c, eb, ef, tm, ehc)
}
\arguments{
\item{temp_k}{temperature in degrees Kelvin}

\item{c}{potential reaction rate}

\item{eb}{baseline energy needed for the reaction to occur (eV)}

\item{ef}{temperature dependence of folding the enzymes used in the metabolic reaction, relative to the melting temperature}

\item{tm}{melting temperature in degrees Kelvin}

\item{ehc}{temperature dependence of the heat capacity between the folded and unfolded state of the enzymes, relative to the melting temperature}
}
\description{
deLong enzyme-assisted Arrhenius model
}
\references{
DeLong, John P., et al. "The combined effects of reactant kinetics and enzyme stability explain the temperature dependence of metabolic rates." Ecology and evolution 7.11 (2017): 3940-3950.
}
\author{
Daniel Padfield
}
