% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_params.R
\name{est_params}
\alias{est_params}
\title{Estimate extra parameters of a thermal performance curve}
\usage{
est_params(model)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}
}
\description{
Estimate extra parameters of a thermal performance curve
}
\details{
Currently estimates:
\itemize{
\item maximum rate (rmax)
\item optimum temperature (topt)
\item critical thermal maximum (ctmax)
\item critical thermal minimum (ctmin)
\item activation energy (e)
\item deactivation energy (eh)
\item q10 value
\item thermal safety margin
\item thermal tolerance
\item skewness
}

If any parameters cannot be calculated for a thermal performance curve, they will return NA.
}
