% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eubank_1973.R
\name{eubank_1973}
\alias{eubank_1973}
\title{Eubank model for fitting thermal performance curves}
\usage{
eubank_1973(temp, topt, a, b)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{topt}{optimum temperature (ºC)}

\item{a}{scale parameter defining the height of the curve}

\item{b}{shape parameter of the curve}
}
\value{
a numeric vector of rate values based on the temperatures and parameter values provided to the function
}
\description{
Eubank model for fitting thermal performance curves
}
\details{
Equation:
\deqn{rate = \frac{a}{(T-T_{\text{opt}})^2+b}}{%
rate = a/((temp-topt)^2)+b}

Start values in \code{get_start_vals} are derived from the data or sensible values from the literature.

Limits in \code{get_lower_lims} and \code{get_upper_lims} are based on extreme values that are unlikely to occur in ecological settings.
}
\note{
Generally we found this model easy to fit.
}
\examples{
# load in ggplot
library(ggplot2)

# subset for the first TPC curve
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'eubank_1973')
# fit model
mod <- nls.multstart::nls_multstart(rate~eubank_1973(temp = temp, topt, a, b),
data = d,
iter = 200,
start_lower = start_vals - 10,
start_upper = start_vals + 10,
lower = get_lower_lims(d$temp, d$rate, model_name = 'eubank_1973'),
upper = get_upper_lims(d$temp, d$rate, model_name = 'eubank_1973'),
supp_errors = 'Y',
convergence_count = FALSE)

# look at model fit
summary(mod)

# get predictions
preds <- data.frame(temp = seq(min(d$temp), max(d$temp), length.out = 100))
preds <- broom::augment(mod, newdata = preds)

# plot
ggplot(preds) +
geom_point(aes(temp, rate), d) +
geom_line(aes(temp, .fitted), col = 'blue') +
theme_bw()

}
\references{
Eubank, W. P., Atmar, J. W. & Ellington, J. J. The significance and thermodynamics of fluctuating versus static thermal environments on Heliothis zea egg development rates. Environ. Entomol. 2, 491–496 (1973).
}
\author{
Francis Windram
}
