% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flinn_1991.R
\name{flinn_1991}
\alias{flinn_1991}
\title{Flinn 1991 model for fitting thermal performance curves}
\usage{
flinn_1991(temp, a, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{parameter that defines the height}

\item{b}{parameter that describes the slope of the increase}

\item{c}{parameter that describes the position and steepness of the decline}
}
\description{
Flinn 1991 model for fitting thermal performance curves
}
\examples{
\dontrun{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'flinn_1991')
mod <- minpack.lm::nlsLM(rate~flinn_1991(temp = temp, a, b, c),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)}
}
\references{
Flinn PW Temperature-dependent functional response of the parasitoid Cephalonomia waterstoni (Gahan) (Hymenoptera, Bethylidae) attacking rusty grain beetle larvae (Coleoptera, Cucujidae). Environmental Entomology, 20, 872–876, (1991)
}
\author{
Daniel Padfield
}
