% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_1987.R
\name{gaussian_1987}
\alias{gaussian_1987}
\title{Gaussian function for TPC}
\usage{
gaussian_1987(temp, rmax, topt, a)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{rmax}{maximum rate at optimum temperature}

\item{topt}{optimum temperature}

\item{a}{related to full curve width}
}
\description{
Gaussian function for TPC
}
\references{
Lynch, M., Gabriel, W., 1987. Environmental tolerance. Am. Nat. 129, 283–303.
}
\author{
Daniel Padfield
}
