% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_breadth.R
\name{get_breadth}
\alias{get_breadth}
\title{Estimate thermal performance breadth of a thermal performance curve}
\usage{
get_breadth(model, level = 0.8)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}

\item{level}{proportion of maximum rate over which thermal performance breadth is calculated}
}
\value{
Numeric estimate of thermal performance breadth (in ºC)
}
\description{
Estimate thermal performance breadth of a thermal performance curve
}
\details{
Thermal performance breadth is calculated as the range of temperatures over which a curve's rate is at least 0.8 of peak. This defaults to a proportion of 0.8 but can be changed using the \code{level} argument.
}
