% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctmax.R
\name{get_ctmax}
\alias{get_ctmax}
\title{Estimate the critical thermal maximum}
\usage{
get_ctmax(model)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}
}
\description{
Estimates the critical thermal maximum from a model model of a thermal performance curve.
}
\details{
Optimum temperature is calculated by predicting over a temperature range 50 degrees beyond the maximum value in the dataset. The value closest to 0 is then extracted. When this is impossible due to the curve formula (i.e the Sharpe-Schoolfield models), the temperature where the rate is 5% of the maximum rate is estimated. Predictions are done every 0.001 temperature value so the estimate of the crticial thermal maximum should be accurate up to 0.001 ºC/K
}
\author{
Daniel Padfield
}
