% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eh.R
\name{get_eh}
\alias{get_eh}
\title{Estimate the deactivation energy of a thermal performance curve}
\usage{
get_eh(model)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}
}
\value{
Numeric estimate of activation energy (eV)
}
\description{
Estimate the deactivation energy of a thermal performance curve
}
\details{
Fits a modified-Boltzmann equation to all raw data beyond the optimum temperature (ºC; as estimated by \code{get_topt}).
}
