% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lower_lims.R
\name{get_lower_lims}
\alias{get_lower_lims}
\title{Set uninformed lower limits on parameter values}
\usage{
get_lower_lims(x, y, model_name)
}
\arguments{
\item{x}{vector of temperature values}

\item{y}{vector of rate values}

\item{model_name}{the name of the model being fitted}
}
\description{
Sets wide lower limits on parameter values for each TPC model
}
\author{
Daniel Padfield
}
