% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_names.R
\name{get_model_names}
\alias{get_model_names}
\title{Lists the models available in rTPC}
\usage{
get_model_names()
}
\value{
character vector of thermal performance curves available in rTPC
}
\description{
Lists the models available in rTPC
}
\examples{
get_model_names()
}
