% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_q10.R
\name{get_q10}
\alias{get_q10}
\title{Estimate the Q10 value of a TPC}
\usage{
get_q10(model)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}
}
\description{
Estimates the Q10 value of a given thermal performance curve
}
\author{
Daniel Padfield
}
