% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skewness.R
\name{get_skewness}
\alias{get_skewness}
\title{Estimates skewness of a thermal performance curve}
\usage{
get_skewness(model)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}
}
\value{
Numeric estimate of skewness
}
\description{
Estimates skewness of a thermal performance curve
}
\details{
Skewness is calculated from the values of activation energy (e) and deactivation energy (eh) as: skewness = e - eh. A negative skewness indicates the TPC is left skewed, the drop after the optimum is steeper than the rise up to the optimum. A positive skewness means that the TPC is right skewed and a value of 0 would mean the curve is symmetrical around the optimum.
}
\concept{params}
