% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_vals.R
\name{get_start_vals}
\alias{get_start_vals}
\title{Estimate start values for TPC fitting}
\usage{
get_start_vals(x, y, model_name)
}
\arguments{
\item{x}{vector of temperature values}

\item{y}{vector of rate values}

\item{model_name}{the name of the model being fitted}
}
\description{
Estimates sensible start values for fitting thermal performance curves
}
\author{
Daniel Padfield
}
