% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thermaltolerance.R
\name{get_thermaltolerance}
\alias{get_thermaltolerance}
\title{Estimate thermal tolerance of a thermal performance curve}
\usage{
get_thermaltolerance(model)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}
}
\value{
Thermal tolerance (in ºC)
}
\description{
Estimate thermal tolerance of a thermal performance curve
}
\details{
Thermal tolerance is calculated as: CTmax - CTmin. This is calculated using the functions \code{get_ctmax} and \code{get_ctmin}.
}
\concept{params}
