% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_topt.R
\name{get_topt}
\alias{get_topt}
\title{Estimate optimum temperature}
\usage{
get_topt(model)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}
}
\description{
Estimates optimum temperature from a model model of a thermal performance curve.
}
\details{
Optimum temperature is calculated by predicting over the temperature range using the previously estimated parameters and subsetting for the largest rate value. Predictions are done every 0.001 temperature value so the estimate of optimum temperature should be accurate up to 0.001 ºC/K
}
\author{
Daniel Padfield
}
