% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tpc_as_formula.R
\name{get_tpc_as_formula}
\alias{get_tpc_as_formula}
\title{Get a formula object for calling a TPC}
\usage{
get_tpc_as_formula(model_name, temp, trait)
}
\arguments{
\item{model_name}{the name of the model being fitted}

\item{temp}{the name of the temperature column}

\item{trait}{the name of the trait column}
}
\value{
A formula calling the expected TPC
}
\description{
Get a formula object for calling a TPC
}
\examples{
get_tpc_as_formula("briere1_1999", "temperature", "rate")
# > rate ~ briere1_1999(temp = temperature, tmin, tmax, a)

}
\author{
Francis Windram
}
\concept{helper}
