% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upper_lims.R
\name{get_upper_lims}
\alias{get_upper_lims}
\title{Set broad upper limits on parameter values}
\usage{
get_upper_lims(x, y, model_name)
}
\arguments{
\item{x}{vector of temperature values}

\item{y}{vector of rate values}

\item{model_name}{the name of the model being fitted}
}
\value{
Named list of upper limits given the data and model being fitted
}
\description{
Sets wide upper limits on parameter values for each TPC model
}
\author{
Daniel Padfield

Francis Windram
}
