% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/johnsonlewin_1946.R
\name{johnsonlewin_1946}
\alias{johnsonlewin_1946}
\title{Johnson-Lewin model}
\usage{
johnsonlewin_1946(temp_k, r0, e, eh, topt)
}
\arguments{
\item{temp_k}{temperature in degrees Kelvin}

\item{r0}{scaling parameter}

\item{e}{activation energy (eV)}

\item{eh}{high temperature de-activation energy (eV)}

\item{topt}{optimum temperature (K)}
}
\description{
Johnson-Lewin model
}
\references{
Johnson, Frank H., and Isaac Lewin. "The growth rate of E. coli in relation to temperature, quinine and coenzyme." Journal of Cellular and Comparative Physiology 28.1 (1946): 47-75.
}
\author{
Daniel Padfield
}
