% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lactin2_1995.R
\name{lactin2_1995}
\alias{lactin2_1995}
\title{Lactin2 function}
\usage{
lactin2_1995(temp, a, b, tmax, delta_t)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{constant that determines the steepness of the rising portion of the curve}

\item{b}{constant that determines the height of the overall curve}

\item{tmax}{changes the temperature at which the curve begins to decelerate beyond the optimum}

\item{delta_t}{thermal safety margin}
}
\description{
Lactin2 function
}
\examples{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'lactin2_1995')
mod <- minpack.lm::nlsLM(rate~lactin2_1995(temp = temp, a, b, tmax, delta_t),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)
}
\references{
Lactin, D.J., Holliday, N.J., Johnson, D.L. & Craigen, R. Improved rate models of temperature-dependent development by arthropods. Environ. Entomol. 24, 69-75 (1995)
}
\author{
Daniel Padfield
}
