% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedgaussian_2006.R
\name{modifiedgaussian_2006}
\alias{modifiedgaussian_2006}
\title{Modified gaussian function for TPC}
\usage{
modifiedgaussian_2006(temp, rmax, topt, a, b)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{rmax}{maximum rate at optimum temperature}

\item{topt}{optimum temperature}

\item{a}{related to full curve width}

\item{b}{allows for asymmetry in the curve fit}
}
\description{
Modified gaussian function for TPC
}
\examples{
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'modifiedgaussian_2006')
mod <- minpack.lm::nlsLM(rate~modifiedgaussian_2006(temp = temp, rmax, topt, a, b),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 1000))

# look at model
summary(mod)
est_params(mod)
}
\author{
Daniel Padfield
}
