% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molnar_2013.R
\name{molnar_2013}
\alias{molnar_2013}
\title{Molnar model for mortality rate}
\usage{
molnar_2013(temp_k, r_tref, e, el, tl, eh, th, tref)
}
\arguments{
\item{temp_k}{temperature in degrees Kelvin}

\item{r_tref}{rate at the standardised temperature, tref}

\item{e}{activation energy (eV)}

\item{el}{low temperature de-activation energy (eV)}

\item{tl}{temperature at which enzyme is 1/2 active and 1/2 suppressed due to low temperatures}

\item{eh}{high temperature de-activation energy (eV)}

\item{th}{temperature at which enzyme is 1/2 active and 1/2 suppressed due to high temperatures}

\item{tref}{standardisation temperature in degrees centigrade}
}
\description{
Molnar model for mortality rate
}
\references{
Molnár, P. K., Kutz, S. J., Hoar, B. M., Dobson, A. P., Metabolic approaches to understanding climate change impacts on seasonal host-macroparasite dynamics. Ecology Letters. 16, 9–21 (2013)
}
