% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneill_1972.R
\name{oneill_1972}
\alias{oneill_1972}
\title{O'Neill 1972 function}
\usage{
oneill_1972(temp, rmax, ctmax, topt, q10)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{rmax}{the rate at optimum temperature}

\item{ctmax}{high temperature at which rates become negative}

\item{topt}{optimum temperature}

\item{q10}{defines the fold change in performance as a result of increasing the temperature by 10 ºC}
}
\description{
O'Neill 1972 function
}
\references{
O’Neill, R.V., Goldstein, R.A., Shugart, H.H., Mankin, J.B. Terrestrial Ecosystem Energy Model. Eastern Deciduous Forest Biome Memo Report Oak Ridge ,. The Environmental Sciences Division of the Oak Ridge National Laboratory. (1972)
}
\author{
Daniel Padfield
}
