% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_2008.R
\name{quadratic_2008}
\alias{quadratic_2008}
\title{Quadratic model}
\usage{
quadratic_2008(temp, a, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{parameter that defines the rate at 0 ºC}

\item{b}{parameter with no biological meaning}

\item{c}{parameter with no biological meaning}
}
\description{
Quadratic model
}
\examples{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'quadratic_2008')
mod <- minpack.lm::nlsLM(rate~quadratic_2008(temp = temp, a, b, c),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)
}
\references{
Montagnes, David JS, et al. "Short‐term temperature change may impact freshwater carbon flux: a microbial perspective." Global Change Biology 14.12: 2823-2838. (2008)
}
\author{
Daniel Padfield
}
