% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratkowsky_1983.R
\name{ratkowsky_1983}
\alias{ratkowsky_1983}
\title{Ratkowsky 1983 model}
\usage{
ratkowsky_1983(temp, tmin, tmax, a, b)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{tmin}{low temperature at which rates become negative}

\item{tmax}{high temperature at which rates become negative}

\item{a}{parameter defined as sqrt(rate)/(temp - tmin)}

\item{b}{empirical parameter needed to fit the data for temperature above topt}
}
\description{
Ratkowsky 1983 model
}
\references{
Ratkowsky, D.A., Lowry, R.K., McMeekin, T.A., Stokes, A.N., Chandler, R.E., Model for bacterial growth rate throughout the entire biokinetic temperature range. J. Bacteriol. 154: 1222–1226 (1983)
}
\author{
Daniel Padfield
}
