% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpeschoolfull_1981.R
\name{sharpeschoolfull_1981}
\alias{sharpeschoolfull_1981}
\title{Full Sharpe-Schoolfield model for fitting thermal performance curves}
\usage{
sharpeschoolfull_1981(temp, r_tref, e, el, tl, eh, th, tref)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{r_tref}{rate at the standardised temperature, tref}

\item{e}{activation energy (eV)}

\item{el}{low temperature de-activation energy (eV)}

\item{tl}{temperature (ºC) at which enzyme is 1/2 active and 1/2 suppressed due to low temperatures}

\item{eh}{high temperature de-activation energy (eV)}

\item{th}{temperature (ºC) at which enzyme is 1/2 active and 1/2 suppressed due to high temperatures}

\item{tref}{standardisation temperature in degrees centigrade. Temperature at which rates are not inactivated by either high or low temperatures}
}
\description{
Full Sharpe-Schoolfield model for fitting thermal performance curves
}
\details{
Equation:
\deqn{rate= \frac{r_{tref} \cdot exp^{\frac{-e}{k} (\frac{1}{temp + 273.15}-\frac{1}{t_{ref}})}}{1+ exp^{\frac{e_l}{k}(\frac{1}{t_l} - \frac{1}{temp + 273.15})} + exp^{\frac{e_h}{k}(\frac{1}{t_h}-\frac{1}{temp + 273.15})}}}{%
rate = r_tref.exp(e/k.(1/tref - 1/(temp + 273.15))) / (1 + exp(-el/k.(1/(tl + 273.15) - 1/(temp + 273.15))) + exp(eh/k.(1/(th + 273.15) - 1/(temp + 273.15))))}
where k is Boltzmann's constant with a value of 8.62e-05
}
\examples{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'sharpeschoolfull_1981')
mod <- minpack.lm::nlsLM(rate~sharpeschoolfull_1981(temp = temp, r_tref, e,
el, tl, eh, th, tref=15),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)
}
\references{
Schoolfield, R. M., Sharpe, P. J. & Magnuson, C. E. Non-linear regression of biological temperature-dependent rate models based on absolute reaction-rate theory. Journal of Theoretical Biology 88, 719-731 (1981)
}
\author{
Daniel Padfield
}
