% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpeschoolhigh_1981.R
\name{sharpeschoolhigh_1981}
\alias{sharpeschoolhigh_1981}
\title{Sharpe Schoolfield Model for high temperature inactivation}
\usage{
sharpeschoolhigh_1981(temp, r_tref, e, eh, th, tref)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{r_tref}{rate at the standardised temperature, tref}

\item{e}{activation energy (eV)}

\item{eh}{high temperature de-activation energy (eV)}

\item{th}{temperature at which enzyme is 1/2 active and 1/2 suppressed due to high temperatures}

\item{tref}{standardisation temperature in degrees centigrade}
}
\description{
Sharpe Schoolfield Model for high temperature inactivation
}
\examples{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'sharpeschoolhigh_1981')
mod <- minpack.lm::nlsLM(rate~sharpeschoolhigh_1981(temp = temp, r_tref, e, eh, th, tref=15),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)
}
\references{
Schoolfield, R. M., Sharpe, P. J. & Magnuson, C. E. Non-linear regression of biological temperature-dependent rate models based on absolute reaction-rate theory. J. Theor. Biol. 88, 719-731 (1981)
}
\author{
Daniel Padfield
}
