% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpeschoollow_1981.R
\name{sharpeschoollow_1981}
\alias{sharpeschoollow_1981}
\title{Sharpe Schoolfield Model for low temperature inactivation}
\usage{
sharpeschoollow_1981(temp, r_tref, e, el, tl, tref)
}
\arguments{
\item{r_tref}{rate at the standardised temperature, tref}

\item{e}{activation energy (eV)}

\item{el}{low temperature de-activation energy (eV)}

\item{tl}{temperature at which enzyme is 1/2 active and 1/2 suppressed due to low temperatures}

\item{tref}{standardisation temperature in degrees centigrade. Temperature at which rates are not inactivated due to low temperature}

\item{(temp}{+ 273.15) temperature in degrees centigrade}
}
\description{
Sharpe Schoolfield Model for low temperature inactivation
}
\references{
Schoolfield, R. M., Sharpe, P. J. & Magnuson, C. E. Non-linear regression of biological temperature-dependent rate models based on absolute reaction-rate theory. J. Theor. Biol. 88, 719-731 (1981)
}
\author{
Daniel Padfield
}
