% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spain_1982.R
\name{spain_1982}
\alias{spain_1982}
\title{Spain function}
\usage{
spain_1982(temp, a, b, c, r0)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{constant that determines the steepness of the rising portion of the curve}

\item{b}{constant that determines the position of topt}

\item{c}{constant that determines the steepness of the decreasing part of the curve}

\item{r0}{the apparent rate at 0 ºC}
}
\description{
Spain function
}
\references{
BASIC Microcomputer Models in Biology. Addison-Wesley, Reading, MA. 1982
}
\author{
Daniel Padfield
}
