% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylorsexton_1972.R
\name{taylorsexton_1972}
\alias{taylorsexton_1972}
\title{Taylor-Sexton model for fitting thermal performance curves}
\usage{
taylorsexton_1972(temp, rmax, tmin, topt)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{rmax}{maximum performance/value of the trait}

\item{tmin}{low temperature (ºC) at which rates become negative}

\item{topt}{optimum temperature (ºC)}
}
\value{
a numeric vector of rate values based on the temperatures and parameter values provided to the function
}
\description{
Taylor-Sexton model for fitting thermal performance curves
}
\details{
Equation:
\deqn{rate = R_{\text{max}} \cdot \frac{-(T-T_{\text{min}})^4 + 2 \cdot (T - T_{\text{min}})^2 \cdot (T_{\text{opt}}-T_{\text{min}})^2}{(T_{\text{opt}}-T_{\text{min}})^4}}{%
rate = Rmax . ((-(T-Tmin)^4 + 2(T-Tmin)^2 . (Topt-Tmin)^2)/(Topt-Tmin)^4)}

Start values in \code{get_start_vals} are derived from the data or sensible values from the literature.

Limits in \code{get_lower_lims} and \code{get_upper_lims} are based on extreme values that are unlikely to occur in ecological settings.
}
\note{
Generally we found this model easy to fit.
}
\examples{
# load in ggplot
library(ggplot2)

# subset for the first TPC curve
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'taylorsexton_1972')
# fit model
mod <- nls.multstart::nls_multstart(rate~taylorsexton_1972(temp = temp, rmax, tmin, topt),
data = d,
iter = 200,
start_lower = start_vals - 10,
start_upper = start_vals + 10,
lower = get_lower_lims(d$temp, d$rate, model_name = 'taylorsexton_1972'),
upper = get_upper_lims(d$temp, d$rate, model_name = 'taylorsexton_1972'),
supp_errors = 'Y',
convergence_count = FALSE)

# look at model fit
summary(mod)

# get predictions
preds <- data.frame(temp = seq(min(d$temp), max(d$temp), length.out = 100))
preds <- broom::augment(mod, newdata = preds)

# plot
ggplot(preds) +
geom_point(aes(temp, rate), d) +
geom_line(aes(temp, .fitted), col = 'blue') +
theme_bw()

}
\references{
Taylor, S. E. & Sexton, O. J. Some implications of leaf tearing in Musaceae. Ecology 53, 143–149 (1972).
}
\author{
Francis Windram
}
\concept{model}
