% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thomas_2012.R
\name{thomas_2012}
\alias{thomas_2012}
\title{Thomas (2012) model for TPC}
\usage{
thomas_2012(temp, a, b, c, topt)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{arbitrary constant}

\item{b}{arbitrary constant}

\item{c}{the range of temperatures over which growth rate is positive, or the thermal niche width (ºC)}

\item{topt}{determines the location of the maximum of the quadratic portion of this function. When b = 0, tref would equal topt}
}
\description{
Thomas (2012) model for TPC
}
\references{
Thomas, Mridul K., et al. A global pattern of thermal adaptation in marine phytoplankton. Science 338.6110, 1085-1088 (2012)
}
\author{
Daniel Padfield
}
