% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thomas_2017.R
\name{thomas_2017}
\alias{thomas_2017}
\title{Thomas model (2017) for fitting thermal performance curves}
\usage{
thomas_2017(temp, a, b, c, d, e)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{birth rate at 0 ºC}

\item{b}{describes the exponential increase in birth rate with increasing temperature}

\item{c}{temperature-independent mortality term}

\item{d}{along with e controls the exponential increase in mortality rates with temperature}

\item{e}{along with d controls the exponential increase in mortality rates with temperature}
}
\description{
Thomas model (2017) for fitting thermal performance curves
}
\details{
Equation:
\deqn{rate = a \cdot exp^{b \cdot temp} - (c + d \cdot exp^{e \cdot temp})}{%
rate = a . exp(b . temp) - (c + d.(exp(e.temp)))}

Start values in \code{get_start_vals} are derived from the data.

Limits in \code{get_lower_lims} and \code{get_upper_lims} are derived from the data or based on extreme values that are unlikely to occur in ecological settings.
}
\note{
Generally we found this model easy to fit.
}
\examples{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'thomas_2017')
mod <- minpack.lm::nlsLM(rate~thomas_2017(temp = temp, a, b, c, d, e),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)
}
\references{
Thomas, Mridul K., et al. Temperature–nutrient interactions exacerbate sensitivity to warming in phytoplankton. Global change biology 23.8 (2017): 3269-3280.
}
