% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_1995.R
\name{weibull_1995}
\alias{weibull_1995}
\title{Weibull model}
\usage{
weibull_1995(temp, a, topt, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{scale the height of the curve}

\item{topt}{optimum temperature}

\item{b}{defines the breadth of the curve}

\item{c}{defines the curve shape}
}
\description{
Weibull model
}
\references{
Angilletta Jr, Michael J. "Estimating and comparing thermal performance curves." Journal of Thermal Biology 31.7 (2006): 541-545.
}
\author{
Daniel Padfield

Jeremy Cohen
}
