% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boatman_2017.R
\name{boatman_2017}
\alias{boatman_2017}
\title{Boatman 2017 function}
\usage{
boatman_2017(temp, rmax, tmin, tmax, a, b)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{rmax}{the rate at optimum temperature}

\item{tmin}{low temperature at which rates become negative}

\item{tmax}{high temperature at which rates become negative}

\item{a}{shape parameter to adjust the asymmetry of the skewness}

\item{b}{is a shape determining parameter, which alters the kurtosis}
}
\description{
Boatman 2017 function
}
\references{
Boatman, T. G., Lawson, T., & Geider, R. J. A key marine diazotroph in a changing ocean: The interacting effects of temperature, CO2 and light on the growth of Trichodesmium erythraeum IMS101. PLoS ONE, 12, e0168796 (2017)
}
\author{
Daniel Padfield
}
