% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briere2_1999.R
\name{briere2_1999}
\alias{briere2_1999}
\title{Briere2 function}
\usage{
briere2_1999(temp, tmin, tmax, a, b)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{tmin}{low temperature at which rates become negative}

\item{tmax}{high temperature at which rates become negative}

\item{a}{scale parameter to adjust maximum rate of the curve}

\item{b}{shape parameter to adjust the asymmetry of the curve}
}
\description{
Briere2 function
}
\references{
Brière, J.F., Pracros, P., Le Roux, A.Y., Pierre, J.S.,  A novel rate model of temperature-dependent development for arthro- pods. Environ. Entomol. 28, 22–29 (1999)
}
\author{
Daniel Padfield
}
