% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_params.R
\name{est_params}
\alias{est_params}
\title{Estimate extra parameters}
\usage{
est_params(model)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}
}
\description{
Estimates extra parameters that are commonly calculated from TPCs
}
\details{
Currently calculates maximum rate (rmax), optimum temperature (topt), critical thermal maximum (ctmax) and minimum (ctmin), thermal tolerance and thermal safety margin. If it cannot be calculated will return NA
}
\author{
Daniel Padfield
}
