% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flinn_1991.R
\name{flinn_1991}
\alias{flinn_1991}
\title{Flinn model for fitting thermal performance curves}
\usage{
flinn_1991(temp, a, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{parameter that controls the height of the curve}

\item{b}{parameter that controls the slope of the initial increase of the curve}

\item{c}{parameter that controls the position and steepness of the decline of the curve}
}
\description{
Flinn model for fitting thermal performance curves
}
\details{
Equation:
\deqn{rate= \frac{1}{1+a+b \cdot temp+c \cdot temp^2}}{%
rate = 1 / (1 + a + b.temp + c.temp^2)}
}
\examples{
\dontrun{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'flinn_1991')
mod <- minpack.lm::nlsLM(rate~flinn_1991(temp = temp, a, b, c),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)}
}
\references{
Flinn PW Temperature-dependent functional response of the parasitoid Cephalonomia waterstoni (Gahan) (Hymenoptera, Bethylidae) attacking rusty grain beetle larvae (Coleoptera, Cucujidae). Environmental Entomology, 20, 872–876, (1991)
}
\author{
Daniel Padfield
}
