% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flinn_1991.R
\name{flinn_1991}
\alias{flinn_1991}
\title{Flinn 1991 model}
\usage{
flinn_1991(temp, a, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{parameter that defines the height}

\item{b}{parameter that describes the slope of the increase}

\item{c}{parameter that describes the position and steepness of the decline}
}
\description{
Flinn 1991 model
}
\references{
Flinn PW Temperature-dependent functional response of the parasitoid Cephalonomia waterstoni (Gahan) (Hymenoptera, Bethylidae) attacking rusty grain beetle larvae (Coleoptera, Cucujidae). Environmental Entomology, 20, 872–876, (1991)
}
\author{
Daniel Padfield
}
