% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_1987.R
\name{gaussian_1987}
\alias{gaussian_1987}
\title{Gaussian model for fitting thermal performance curves}
\usage{
gaussian_1987(temp, rmax, topt, a)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{rmax}{maximum rate at optimum temperature}

\item{topt}{optimum temperature (ºC)}

\item{a}{related to the full curve width}
}
\description{
Gaussian model for fitting thermal performance curves
}
\details{
Equation:
\deqn{rate = r_{max} \cdot exp^{\bigg(-0.5 \left(\frac{|temp-t_{opt}|}{a}\right)^2\bigg)}}{%
rate = rmax.exp(-0.5.(abs(temp - topt)/a)^2)}

Start values in \code{get_start_vals} are derived from the data

Limits in \code{get_lower_lims} and \code{get_upper_lims} are based on extreme values that are unlikely to occur in ecological settings.
}
\note{
Generally we found this model easy to fit.
}
\examples{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'gaussian_1987')
mod <- minpack.lm::nlsLM(rate~gaussian_1987(temp = temp, rmax, topt, a),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)
}
\references{
Lynch, M., Gabriel, W., Environmental tolerance. The American Naturalist. 129, 283–303. (1987)
}
