% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_e.R
\name{get_e}
\alias{get_e}
\title{Estimate the activation energy}
\usage{
get_e(model)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}
}
\description{
Estimates the activation energy of a given thermal performance curve
}
\author{
Daniel Padfield
}
