% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eh.R
\name{get_eh}
\alias{get_eh}
\title{Estimate the deactivation energy}
\usage{
get_eh(model)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}
}
\description{
Estimates the deactivation energy of a given thermal performance curve
}
\author{
Daniel Padfield
}
