% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_q10.R
\name{get_q10}
\alias{get_q10}
\title{Estimate the q10 value of a thermal performance curve}
\usage{
get_q10(model)
}
\arguments{
\item{model}{nls model object that contains a model of a thermal performance curve}
}
\value{
q10 value
}
\description{
Estimate the q10 value of a thermal performance curve
}
\details{
Fits the q10 portion of \code{rezende_2019} to all raw data below the optimum temperature (ºC; as estimated by \code{get_topt}).
}
