% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinshelwood_1947.R
\name{hinshelwood_1947}
\alias{hinshelwood_1947}
\title{Hinshelwood 1947 model}
\usage{
hinshelwood_1947(temp, a, e, b, eh)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{pre-exponential factor}

\item{e}{activation energy (eV)}

\item{b}{pre-exponential factor}

\item{eh}{de-activation energy (eV)}
}
\description{
Hinshelwood 1947 model
}
\references{
Hinshelwood C.N. The Chemical Kinetics of the Bacterial Cell. Oxford University Press. (1947)
}
\author{
Daniel Padfield
}
