% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinshelwood_1947.R
\name{hinshelwood_1947}
\alias{hinshelwood_1947}
\title{Hinshelwood 1947 model}
\usage{
hinshelwood_1947(temp, a, e, b, eh)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{pre-exponential factor}

\item{e}{activation energy (eV)}

\item{b}{pre-exponential factor}

\item{eh}{de-activation energy (eV)}
}
\description{
Hinshelwood 1947 model
}
\examples{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'hinshelwood_1947')
mod <- minpack.lm::nlsLM(rate~hinshelwood_1947(temp = temp, a, e, b, eh),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)
}
\references{
Hinshelwood C.N. The Chemical Kinetics of the Bacterial Cell. Oxford University Press. (1947)
}
\author{
Daniel Padfield
}
