% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joehnk_2008.R
\name{joehnk_2008}
\alias{joehnk_2008}
\title{Jöhnk 2008 function}
\usage{
joehnk_2008(temp, rmax, topt, a, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{rmax}{the rate at optimum temperature}

\item{topt}{optimum temperatute}

\item{a}{parameter with no biological meaning}

\item{b}{parameter with no biological meaning}

\item{c}{parameter with no biological meaning}
}
\description{
Jöhnk 2008 function
}
\references{
Joehnk, Klaus D., et al. "Summer heatwaves promote blooms of harmful cyanobacteria." Global change biology 14.3: 495-512 (2008)
}
\author{
Daniel Padfield
}
