% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kamykowski_1985.R
\name{kamykowski_1985}
\alias{kamykowski_1985}
\title{Kamykowski 1985 function}
\usage{
kamykowski_1985(temp, tmin, tmax, a, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{tmin}{low temperature at which rates become negative}

\item{tmax}{high temperature at which rates become negative}

\item{a}{parameter with no biological meaning}

\item{b}{parameter with no biological meaning}

\item{c}{parameter with no biological meaning}
}
\description{
Kamykowski 1985 function
}
\examples{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name == 'kamykowski_1985')
mod <- minpack.lm::nlsLM(rate~kamykowski_1985(temp = temp, tmin, tmax, a, b, c),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)
}
\references{
Kamykowski, Daniel. "A survey of protozoan laboratory temperature studies applied to marine dinoflagellate behavior from a field perspective." Contributions in Marine Science. 1985. (1985).
}
\author{
Daniel Padfield
}
