% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lactin2_1995.R
\name{lactin2_1995}
\alias{lactin2_1995}
\title{Lactin2 function}
\usage{
lactin2_1995(temp, p, c, tmax, delta_t)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{p}{constant that determines the steepness of the rising portion of the curve}

\item{c}{constant that determines the height of the overall curve}

\item{tmax}{changes the temperature at which the curve begins to decelerate beyond the optimum}

\item{delta_t}{thermal safety margin}
}
\description{
Lactin2 function
}
\references{
Lactin, D.J., Holliday, N.J., Johnson, D.L. & Craigen, R. Improved rate models of temperature-dependent development by arthropods. Environ. Entomol. 24, 69-75 (1995)
}
\author{
Daniel Padfield
}
