% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedgaussian_2006.R
\name{modifiedgaussian_2006}
\alias{modifiedgaussian_2006}
\title{Modified gaussian function for TPC}
\usage{
modifiedgaussian_2006(temp, rmax, topt, a, b)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{rmax}{maximum rate at optimum temperature}

\item{topt}{optimum temperature}

\item{a}{related to full curve width}

\item{b}{allows for asymmetry in the curve fit}
}
\description{
Modified gaussian function for TPC
}
\references{

}
\author{
Daniel Padfield
}
