% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_2008.R
\name{quadratic_2008}
\alias{quadratic_2008}
\title{Quadratic model for fitting thermal performance curves}
\usage{
quadratic_2008(temp, a, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{parameter that defines the rate at 0 ºC}

\item{b}{parameter with no biological meaning}

\item{c}{parameter with no biological meaning}
}
\description{
Quadratic model for fitting thermal performance curves
}
\details{
Equation:
\deqn{rate = a + b \cdot temp + c \cdot temp^2}{%
rate = a + b.temp + c.temp^2}

Start values in \code{get_start_vals} are derived from the data using previous methods in the literature

Limits in \code{get_lower_lims} and \code{get_upper_lims} are based on extreme values that are unlikely to occur in ecological settings.
}
\note{
Generally we found this model easy to fit.
}
\examples{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'quadratic_2008')
mod <- minpack.lm::nlsLM(rate~quadratic_2008(temp = temp, a, b, c),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)
}
\references{
Montagnes, David JS, et al. Short‐term temperature change may impact freshwater carbon flux: a microbial perspective. Global Change Biology 14.12: 2823-2838. (2008)
}
