% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic_2008.R
\name{quadratic_2008}
\alias{quadratic_2008}
\title{Quadratic model}
\usage{
quadratic_2008(temp, a, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{parameter that defines the rate at 0 ºC}

\item{b}{parameter with no biological meaning}

\item{c}{parameter with no biological meaning}
}
\description{
Quadratic model
}
\references{
Montagnes, David JS, et al. "Short‐term temperature change may impact freshwater carbon flux: a microbial perspective." Global Change Biology 14.12: 2823-2838. (2008)
}
\author{
Daniel Padfield
}
