% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickfit_tpc.R
\name{quickfit_tpc}
\alias{quickfit_tpc}
\title{Perform a quick tpc fit}
\usage{
quickfit_tpc(
  data,
  model_name,
  temp,
  trait,
  start_adjusts = 0,
  iter = 150,
  lhstype = "none",
  gridstart = FALSE,
  force = FALSE
)
}
\arguments{
\item{data}{the data to fit a model to}

\item{model_name}{the model name as a string}

\item{temp}{the column name (as a string) containing the temperature data}

\item{trait}{the column name (as a string) containing the temperature data}

\item{start_adjusts}{any adjustments to make to the lower and upper starting bounds. If \code{0 < start_adjusts < 1}, this will be interpreted as a proportion of the base starting values.}

\item{iter}{number of combinations of starting parameters which will be tried (as in \code{\link[nls.multstart]{nls_multstart}})}

\item{lhstype}{method to use for Latin Hypercube Sampling using \code{\link[lhs]{lhs}} (as in \code{\link[nls.multstart]{nls_multstart}})}

\item{gridstart}{whether to run a gridstart approach (interpreting iter as the number of samples to take across each parameter, 
so \code{3} will become \code{c(3,3,3)} for a 3-parameter model)}

\item{force}{whether to force a gridstart even with very large numbers of iterations}
}
\value{
The nls model object of the fit model
}
\description{
Performs a simple TPC fit using \code{\link[nls.multstart]{nls_multstart}}. This function tries to use a sensible default configuration, 
however if you need to use the more esoteric elements of \code{\link[nls.multstart]{nls_multstart}} then you will need to construct your own.
}
\examples{
\dontrun{
data("chlorella_tpc")
subs <- subset(chlorella_tpc, curve_id == 1)
quickfit_tpc(subs, "briere1_1999", "temp", "rate")

quickfit_tpc(
  subs,
  "briere1_1999",
  "temp",
  "rate",
  start_adjusts = 10,
  iter = 150,
  lhstype = "maximin"
)

quickfit_tpc(
  subs,
  "briere1_1999",
  "temp",
  "rate",
  start_adjusts = 10,
  iter = 5,
  gridstart = TRUE
)
}

}
\author{
Francis Windram
}
\concept{helper}
