% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rezende_2019.R
\name{rezende_2019}
\alias{rezende_2019}
\title{Rezende 2019 model}
\usage{
rezende_2019(temp, q10, a, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{parameter describing shifts in the vertical axis}

\item{b}{parameter threshold temperature beyond which the downward curve starts}

\item{c}{parameter controlling the rate of decline beyond the threshold temperature, b.}
}
\description{
Rezende 2019 model
}
\references{
Rezende, Enrico L., and Francisco Bozinovic. "Thermal performance across levels of biological organization." Philosophical Transactions of the Royal Society B 374.1778 (2019): 20180549.
}
\author{
Daniel Padfield
}
