% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpeschoolhigh_1981.R
\name{sharpeschoolhigh_1981}
\alias{sharpeschoolhigh_1981}
\title{Sharpe Schoolfield Model for high temperature inactivation}
\usage{
sharpeschoolhigh_1981(temp_k, r_tref, e, eh, th, tref)
}
\arguments{
\item{temp_k}{temperature in degrees Kelvin}

\item{r_tref}{rate at the standardised temperature, tref}

\item{e}{activation energy (eV)}

\item{eh}{high temperature de-activation energy (eV)}

\item{th}{temperature at which enzyme is 1/2 active and 1/2 suppressed due to high temperatures}

\item{tref}{standardisation temperature in degrees centigrade}
}
\description{
Sharpe Schoolfield Model for high temperature inactivation
}
\references{
Schoolfield, R. M., Sharpe, P. J. & Magnuson, C. E. Non-linear regression of biological temperature-dependent rate models based on absolute reaction-rate theory. J. Theor. Biol. 88, 719-731 (1981)
}
\author{
Daniel Padfield
}
