% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thomas_2017.R
\name{thomas_2017}
\alias{thomas_2017}
\title{Thomas 2017}
\usage{
thomas_2017(temp, a, b, c, d, e)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{birth rate at 0 ºC}

\item{b}{describes the exponential increase in birth rate with increasing temperature}

\item{c}{temperature-independent mortality term}

\item{d}{along with e controls the exponential increase in mortality rates with temperature}

\item{e}{along with d controls the exponential increase in mortality rates with temperature}
}
\description{
Thomas 2017
}
\references{
Thomas, Mridul K., et al. "Temperature–nutrient interactions exacerbate sensitivity to warming in phytoplankton." Global change biology 23.8 (2017): 3269-3280.
}
\author{
Daniel Padfield
}
