% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_1995.R
\name{weibull_1995}
\alias{weibull_1995}
\title{Weibull model}
\usage{
weibull_1995(temp, a, topt, b, c)
}
\arguments{
\item{temp}{temperature in degrees centigrade}

\item{a}{scale the height of the curve}

\item{topt}{optimum temperature}

\item{b}{defines the breadth of the curve}

\item{c}{defines the curve shape}
}
\description{
Weibull model
}
\examples{
# load in data
data('chlorella_tpc')
d <- subset(chlorella_tpc, curve_id == 1)

# get start values and fit model
start_vals <- get_start_vals(d$temp, d$rate, model_name = 'weibull_1995')
mod <- minpack.lm::nlsLM(rate~weibull_1995(temp = temp, a, topt, b, c),
data = d,
start = start_vals,
control = minpack.lm::nls.lm.control(maxiter = 100))

# look at model
summary(mod)
est_params(mod)
}
\references{
Angilletta Jr, Michael J. "Estimating and comparing thermal performance curves." Journal of Thermal Biology 31.7 (2006): 541-545.
}
\author{
Daniel Padfield

Jeremy Cohen
}
